fix(feature-card): renaming feature link to feature card
