+++ packages/react/src/patterns/blocks/FeatureCard/__stories__/index.scss
@@ -1,0 +1,1 @@
+@import '../../../../../../styles/scss/patterns/blocks/feature-card/index';
