+++ packages/react/src/patterns/blocks/FeatureCard/index.js
@@ -4,4 +4,4 @@
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
-export { default as FeaturedLink } from './FeaturedLink';
+export { default as FeatureCard } from './FeatureCard';
