+++ packages/react/src/patterns/sections/CardSection/CardSection.js
@@ -54,6 +54,7 @@ const CardSection = ({ heading, cards, theme }) => {
 
   /**
    * sets the class name based on theme type
+   *
    * @private
    * @param {string} theme theme type ( g10 | g100 | white/default )
    * @returns {string} theme css class names
@@ -78,16 +79,16 @@ const CardSection = ({ heading, cards, theme }) => {
                     <Card
                       key={index}
                       image={card.image}
-                      heading={card.title}
+                      heading={card.heading}
                       eyebrow={card.eyebrow}
                       copy={card.copy}
                       cta={{
-                        href: card.link.href,
+                        href: card.cta.href,
                         icon: {
                           src: ArrowRight20,
                         },
                       }}
-                      target={card.link.target}
+                      target={card.cta.target}
                       type="link"
                     />
                   </div>
