+++ packages/styles/scss/ibm-dotcom-styles.scss
@@ -30,7 +30,7 @@
 // Patterns - Blocks
 @import 'patterns/blocks/contentgroupcards/contentgroupcards';
 @import 'patterns/blocks/content-group-pictograms/content-group-pictograms';
-@import 'patterns/blocks/FeaturedLink/featuredlink';
+@import 'patterns/blocks/feature-card/feature-card';
 @import 'patterns/blocks/listsection/listsection';
 @import 'patterns/blocks/logo-grid/logo-grid';
 @import 'patterns/blocks/simplelongform/simplelongform';
