--- packages/styles/scss/patterns/blocks/FeaturedLink/README.md
@@ -1,24 +1,0 @@
-### SCSS
-
-#### Usage
-
-Import the package css into the top of your main CSS file.
-
-```css
-@import '@carbon/ibmdotcom-styles/scss/patterns/blocks/featuredlink/index';
-```
-
-#### Elements
-
-| Class                          | Description                                         |
-| ------------------------------ | --------------------------------------------------- |
-| `.bx--content-group__children` | Element containing all featured link group children |
-| `.bx--card`                    | Featured Link card link item                        |
-| `.bx--card__wrapper`           | Card Link wrapper for the Image element             |
-| `.bx--card .bx--image`         | Image element                                       |
-
-#### Blocks
-
-| Class               | Description                                            |
-| ------------------- | ------------------------------------------------------ |
-| `.bx--featuredlink` | Highest level class name for the featured link pattern |
