+++ packages/styles/scss/patterns/blocks/feature-card/README.md
@@ -1,0 +1,21 @@
+### SCSS
+
+#### Usage
+
+Import the package css into the top of your main CSS file.
+
+```css
+@import '@carbon/ibmdotcom-styles/scss/patterns/blocks/feature-card/index';
+```
+
+#### Elements
+
+| Class                     | Description                                 |
+| ------------------------- | ------------------------------------------- |
+| `.bx--feature-card__card` | Card element within the FeatureCard pattern |
+
+#### Blocks
+
+| Class               | Description                                          |
+| ------------------- | ---------------------------------------------------- |
+| `.bx--feature-card` | Highest level class name for the FeatureCard pattern |
