+++ packages/styles/scss/patterns/blocks/feature-card/index.scss
@@ -6,5 +6,5 @@
 //
 
 @import '../../../globals/imports';
-@import 'featuredlink';
+@import 'feature-card';
 @import '../../../components/image/image';
